requirements.txt
aws-cdk.core==1.130.0
aws-cdk.aws-lambda==1.130.0
aws-cdk.aws-apigateway==1.130.0
aws-cdk.aws-iot==1.130.0